Refactor hooks.go to use fileloader and azureloader packages for configuration loading
