+++ upload-server/internal/reporters/reporters.go
@@ -1,0 +1,13 @@
+package reporters
+
+import (
+	"context"
+)
+
+type Identifiable interface {
+	Identifier() string
+}
+
+type Reporter interface {
+	Publish(context.Context, Identifiable) error
+}
